+++ src/main/java/kitchenpos/order/domain/repository/OrderLineItemRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.repository;
 
+import kitchenpos.order.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
