+++ src/main/java/kitchenpos/order/domain/repository/OrderRepository.java
@@ -1,6 +1,8 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.repository;
 
 import java.util.List;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
